pyproject.toml
[tool.poetry]
name = "horoscope-bot"
version = "0.1.0"
description = ""
authors = ["a.kazachinskiy <kazachinskiy.a@romir.ru>"]
readme = "README.md"
packages = [{include = "horoscope_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
aiogram = "^2.25.1"
python-dotenv = "^1.0.0"
h5py = "^3.8.0"
scikit-learn = "^1.2.2"
datasets = "^2.12.0"
evaluate = "^0.4.0"
accelerate = "^0.18.0"
python-docx = "^0.8.11"
selenium = "^4.9.0"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
pandas = "^2.0.1"
transformers = "^4.28.1"
openpyxl = "^3.1.2"
torch = "^2.0.0"
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# PYTEST

[tool.pytest]
addopts = "--cov=meal_backend --cov-report html:coverage-report-html --cov-fail-under=60"
python_files = "test_*.py"
python_classes = ["*Test", "Test*"]
python_functions = "test_*"
testpaths = "tests/"

[tool.coverage.report]
exclude_lines = [
    "# pragma: no cover",
    "def __repr__",
    "def __str__",
    "def __unicode__",
]
show_missing = true
skip_empty = true
omit = [
  "*/__init__.py",
  "*/repositories/*",
  "*/migrations/__main__.py",
  "*/core/logging/*"
]

[tool.coverage.html]
directory = "pytest-cov-report"

[tool.coverage.run]
branch = true
concurrency = ['thread', 'greenlet']

[tool.pytest.ini_options]
asyncio_mode = "auto"

# LINTERS

[tool.pylint.master]
extension-pkg-allow-list = ["pydantic"]
ignore-paths=["matcher.db.migrations"]

[tool.pylint.format]
max-line-length = 120

[tool.pylint.basic]
good-names = ["db"]

[tool.pylint."MESSAGES CONTROL"]
disable = [
  "C0114",  # missing-module-docstring
  "C0116",  # missing-function-docstring
  "R0903",  # too-few-public-methods (for sqlalchemy tables and pydantic classes)
  "C0115",  # missing-class-docstring (for api routes)
  "C0103",  # invalid-name (for pydatic settings)
  "I1101",  # c-extension-no-member
]

[tool.mypy]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
ignore_missing_imports = false
no_implicit_optional = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
sections = ["FUTURE","STDLIB","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"]
profile = "black"
known_local_folder = "horoscoper"
lines_after_imports = 2
multi_line_output = 3
line_length = 120
include_trailing_comma = true
